requirements.txt
aiogram==3.9.0
gspread==6.1.2
google-auth==2.29.0
APScheduler==3.10.4